Use `backoff_retry()` when fetching client config
